+++ resolwe_bio/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'resolwe_bio.apps.BaseConfig'  # pylint: disable=invalid-name